Pull `send` docs to the top
It's a bit of a long scroll down the docs page to find how to actually send off a request. I think this is information that should be at the start of the docs, and it introduces the user to all the necessary types and some functions that they can then investigate further.